+++ src/index.ts
@@ -1,5 +1,4 @@
-import "requestidlecallback-polyfill";
-
+import "./polyfills";
 import maplibregl from "maplibre-gl";
 
 // Types from MapLibre are not re-exported one by one