+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - None
 
 ### 🐛 Bug Fixes
-- None
+- Adds a fix to prevent `requestIdleCallback` iife polyfill from failing on non-browser imports of the SDK.
 
 ### ⚙️ Others
 - None